requirements.txt
beautifulsoup4
Pillow
Flask
dataclasses-json